Move functions to testutils
